+++ cmd/neofs-adm/internal/modules/fschain/container.go
@@ -105,6 +105,8 @@ func dumpContainers(cmd *cobra.Command, _ []string) error {
 			cnt.EACL = ea
 		}
 
+		cnt.ID = id.String()
+
 		containers = append(containers, cnt)
 	}
 
@@ -183,8 +185,14 @@ func restoreContainers(cmd *cobra.Command, _ []string) error {
 
 	var id cid.ID
 	b := smartcontract.NewBuilder()
-	for _, cnt := range containers {
-		id = cid.NewFromMarshalledContainer(cnt.Value)
+	for i, cnt := range containers {
+		if cnt.ID != "" {
+			if err = id.DecodeString(cnt.ID); err != nil {
+				return fmt.Errorf("invalid ID in container #%d: %w", i, err)
+			}
+		} else {
+			id = cid.NewFromMarshalledContainer(cnt.Value)
+		}
 		if _, ok := requested[id]; !ok {
 			continue
 		}
@@ -234,6 +242,7 @@ func restoreContainers(cmd *cobra.Command, _ []string) error {
 
 // Container represents container struct in contract storage.
 type Container struct {
+	ID        string `json:"id"`
 	Value     []byte `json:"value"`
 	Signature []byte `json:"signature"`
 	PublicKey []byte `json:"public_key"`